This PR was merged into the 5.4 branch.

Discussion
----------

[HttpClient][Mailer][Process] always pass microseconds to usleep as integers

| Q             | A
| ------------- | ---
| Branch?       | 5.4
| Bug fix?      | yes
| New feature?  | no
| Deprecations? | no
| Issues        | Fix #52856
| License       | MIT

Commits
-------

fb0ef06872 always pass microseconds to usleep as integers

+++ src/Symfony/Component/HttpClient/Response/TransportResponseTrait.php
@@ -303,7 +303,7 @@ trait TransportResponseTrait
             }
 
             if (-1 === self::select($multi, min($timeoutMin, $timeoutMax - $elapsedTimeout))) {
-                usleep(min(500, 1E6 * $timeoutMin));
+                usleep((int) min(500, 1E6 * $timeoutMin));
             }
 
             $elapsedTimeout = microtime(true) - $lastActivity;
